refactor: Clean up unused budget methods in settings + move packages into budget package
