+++ app/src/main/java/dev/ridill/oar/budgetCycles/presentation/currencyUpdate/CurrencySelectionViewModel.kt
@@ -1,4 +1,4 @@
-package dev.ridill.oar.settings.presentation.currencyUpdate
+package dev.ridill.oar.budgetCycles.presentation.currencyUpdate
 
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.lifecycle.ViewModel
