+++ app/src/main/java/dev/ridill/oar/di/MonthlyBudgetModule.kt
@@ -5,7 +5,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
 import dev.ridill.oar.core.domain.util.EventBus
-import dev.ridill.oar.settings.presentation.budgetUpdate.UpdateBudgetViewModel
+import dev.ridill.oar.budgetCycles.presentation.budgetUpdate.UpdateBudgetViewModel
 
 @Module
 @InstallIn(ViewModelComponent::class)
