+++ app/src/main/java/dev/ridill/oar/settings/domain/repositoty/SettingsRepository.kt
@@ -10,8 +10,6 @@ interface SettingsRepository {
     suspend fun updateAppTheme(theme: AppTheme)
     fun getDynamicColorsEnabled(): Flow<Boolean>
     suspend fun toggleDynamicColors(enabled: Boolean)
-    fun getCurrentBudget(): Flow<Long>
-    suspend fun updateCurrencyPreference(currency: Currency)
     fun getTransactionAutoDetectEnabled(): Flow<Boolean>
     suspend fun toggleAutoDetectTransactions(enabled: Boolean)
     suspend fun getShowTransactionAutoDetectInfoValue(): Boolean
